Refactor looker
Split the looker robot between its definition and service. This way, we
can redeploy the service without messing with the task definition (
including creating a new family! )

Update looker to use an ssm parameter with a secure string storing the
discord token instead of using a secret. This way, we can bring it up
and down pretty easily in aws

Set force_new_deployment in the looker service