Decouple aviary install and its inventory
Fetch the inventory ourselves, separate from aviary's default location
for it. Optionally pass inventory_root, for where our inventory might
not be right at the top of a repo. Pass our inventory to aviary with
--inventory flag. Go back to generating conf values at $inventory, since
it points somewhere else.

Construct the url to install aviary outside of bash, either using a
whole install_url, or a ref on my branch