Configure echo for consul dns
Tell nomad to tell consul to tell dns resolvers how to get to echo. I
_think_ this should enable the `echo.service.consul` dns we've been
setting up

Switching to bridge mode hands ip over to cni, and the connect {} block
tells nomad that we want `echo` to be dns resolveable. To put the
connect block in service we had to pull service block down to group. Not
sure what else this changing, or what the significance of a group vs
task service...